requirements.txt
streamlit
pandas
numpy
matplotlib.pyplot
seaborn
plotly.express
scikit-learn
joblib